refactored new database and changed number types to floats for optimi…

+++ WeatherWorryWonder/Models/ResultView.cs
@@ -8,18 +8,18 @@ namespace WeatherWorryWonder.Models
     public class ResultView
     {
         public string PollutantWarning{ get; set; }
-        public decimal CO { get; set; }
-        public decimal C2H4OPPM { get; set; }
-        public decimal O3AQI { get; set; }
-        public decimal Second03AQI { get; set; }
-        public decimal Third03AQI { get; set;}
-        public decimal PM25AQI { get; set; }
-        public decimal NO2AQI { get; set; }
-        public decimal SO2AQI { get; set; }
-        public decimal PredictedAQITomorrow { get; set; }
-        public decimal PredictedAQI3Day { get; set; }
-        public decimal PredictedAQI5Day { get; set; }
-        public decimal EpaAQI { get; set; }
+        public float CO { get; set; }
+        public float C2H4OPPM { get; set; }
+        public float O3AQI { get; set; }
+        public float Second03AQI { get; set; }
+        public float Third03AQI { get; set;}
+        public float PM25AQI { get; set; }
+        public float NO2AQI { get; set; }
+        public float SO2AQI { get; set; }
+        public float PredictedAQITomorrow { get; set; }
+        public float PredictedAQI3Day { get; set; }
+        public float PredictedAQI5Day { get; set; }
+        public float EpaAQI { get; set; }
         public string SensorName { get; set; }
         public double UserLatitude { get; set; }
         public double UserLongitude { get; set; }
